Initial OS commit, DOES NOT WORK
Removed a bunch of stuff that was not my own work, and removed all
references to text/material/whatever that may be considered a direct
violation of, or facilitation of violating, academic integrity. There
should not be any way for this repo to be found online by searching for
terms relevant to any given course. Due to the amount of removed
material, I am marking this as not working because I haven't tested it
yet and I guarantee that it doesn't work.